saas-smart-report-writer/smart_report_writer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_report_writer"
version = "0.1.0"
description = "Modular Smart Report Writer using LangChain + Ollama"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name="Your Name", email="you@example.com"}]

dependencies = [
  "langchain>=0.2",
  "langchain-core>=0.2",
  "langchain-ollama>=0.1.0",
  "tenacity>=8.2",
  "pandas>=2.0",
  "jinja2>=3.1",
  # optional
  "chardet>=5.2; extra == 'all'",
  "pdfplumber>=0.11; extra == 'all'",
  "python-docx>=1.1; extra == 'all'",
  "docxtpl>=0.16; extra == 'all'",
  "beautifulsoup4>=4.12; extra == 'all'",
  "lxml>=5.2; extra == 'all'",
]

[project.optional-dependencies]
all = ["chardet>=5.2", "pdfplumber>=0.11", "python-docx>=1.1", "docxtpl>=0.16", "beautifulsoup4>=4.12", "lxml>=5.2"]

[project.scripts]
smart-report-writer = "smart_report_writer.cli:main"
